pyproject.toml
[tool.poetry]
name = "microshop-fastapi"
version = "0.1.0"
description = ""
authors = ["Дарья Коннова <dashustic@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standart"], version = "^0.24.0.post1"}
pydantic = {extras = ["email"], version = "^2.5.2"}
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.23"}
aiosqlite = "^0.19.0"
pydantic-settings = "^2.1.0"
alembic = "^1.13.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
